rename types/animal.ts to types/types.ts to save all types in one file
